backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
Flask-RESTful==0.3.9
Flask-JWT==0.3.2
SQLAlchemy==1.3.22
requests==2.23.0
mysql-connector-python==8.0.21
python-dateutil==2.8.2
